refactor(showcase): consolidated landing page showcase items style
