+++ pages/index.tsx
@@ -37,7 +37,7 @@ import { Footer } from 'components/footer'
 import Header from 'components/header'
 import SandpackEmbed from 'components/sandpack-embed'
 import SEO from 'components/seo'
-import ShowcaseSection from 'components/showcase-section'
+import ShowcaseSection from 'components/showcase/showcase-section'
 import TweetCard from 'components/tweet-card'
 import { App, Index } from 'configs/sandpack-contents/homepage/files'
 import tweets from 'configs/tweets.json'
